RESTCONF-ish thingy: let's get it started
There's this Netopeer thingy for NETCONF. In Czech, it's pronounced as
the Czech word for a bat (the flying mouse variety, with all due respect
for bats), so I wanted to find something similar for RESTCONF,
preferably starting with letter "R".

Enter "rousette", or rousettus: https://en.wikipedia.org/wiki/Rousettus,
which is "kaloň" in Czech.

Note that this is specifically not "roussette" (double "s"), that would
be a shark which the Wikipedia calls a "nursehound". While it's got
double-s, which is awesome because it points to sysrepo, it starts with
an "n", which is a no-go, there are too mani things N-related with
NETCONF already.

Change-Id: I9359989a995ce0f4e15539450dbd076cda25a725