CI: get started with some async HTTP2 bits
I had a few back-and-forth moments on how to get that thing installed.
The biggest problem turned out to be a peculiarity of autotools+clang
where the usual way of doing stuff via LD=clang, LDFLAGS=-fsanitize=...
would simply break. I tracked this down to libtool or the linker or
something not creating the expected file `libnghttp2.so.14.20.1`, just
the symlinks.

Since this means that we effectively cannot use sanitizers for this lib,
and that we've already deprecated libc++ as the STL because it's lagging
behind the implementation in GCC, I can as well move that provisioning
thingy to Ansible -- why not, I guess.

Change-Id: Ia59d21b605f82eac8016c6f296694e617d8dac7e