http: log the forwarded header as well
This is an HTTP/2 server only, and HTTP/2 from a browser "requires TLS",
effectively. Since this will very likely run on boxes that are not
publicly available and where cert management could be an issue, it will
be very likely running behind a reverse proxy that translates HTTP/1.1
calls to HTTP/2 in this backend. When that happens, make sure that the
log keeps some unique identification of each connection.

One example of doing that is via:

 nghttpx -f '*,1080;no-tls' -b '::1,10080;;proto=h2' --add-forwarded=for

Change-Id: Iff2c948958bd1ea6c09d6fa9874777bfadfa922d