Generate YANG "push-update" notifications upon changes
It turned out that the YANG push RFC defines a particular format for
the change stream, and that change stream is not available via sysrepo
just yet. However, it's perfectly OK to push *subtrees* of the full DS,
and this patch implements just that.

Whenever there's a change in any manually designated subtree, a
push-update is generated and sent over a text/event-stream. As a result,
we have an asynchronous notification that's compliant with both RESTCONF
and YANG push.

Change-Id: I416822e12108136fbe76ea505e7c9b78b5267e3f
See-also: https://tools.ietf.org/html/rfc8641#section-3.7