Add a hi-res timestamp to each event
I was a bit lazy and this code assumes that when the initial data gets
pushed, it's tagged with the current time. That's not awesome, but hey.

I was not lazy at all and uber-riced the time formatting, though. I
stopped only when I was about to add a compile-time log10
implementation.

Change-Id: I38f56caace38b2671cd0a68a37298a6a38486cd7