Fix race between EventStream::enqueue and response::on_close
When the response's on_close() gets invoked and it returns, the response
should not be used anymore -- that's what the docs say. However, there
was a race here because enqueue() might be already running while
on_close fires, and this is the result:

 Mar 25 00:13:42 line-DB00QCP9 rousette[445]: change: 50095 bytes
 Mar 25 00:13:42 line-DB00QCP9 rousette[445]: [::1]:38104: new event, ∑ queue size = 0
 Mar 25 00:13:42 line-DB00QCP9 rousette[445]: [::1]:38104: sent one event
 Mar 25 00:13:42 line-DB00QCP9 rousette[445]: [::1]:38104: sleeping
 Mar 25 00:13:42 line-DB00QCP9 rousette[445]: change: 50094 bytes
 Mar 25 00:13:42 line-DB00QCP9 rousette[445]: [::1]:38104: new event, ∑ queue size = 0
 Mar 25 00:13:42 line-DB00QCP9 rousette[445]: [::1]:38104: closed (INTERNAL_ERROR)
 Mar 25 00:13:42 line-DB00QCP9 systemd[1]: rousette.service: Main process exited, code=killed, status=11/SEGV

 (gdb) bt
 #0  0xb6ec2b44 in nghttp2::asio_http2::server::stream::handler() const () from /lib/libnghttp2_asio.so.1
 #1  0xb6ec23cc in nghttp2::asio_http2::server::response_impl::resume() () from /lib/libnghttp2_asio.so.1
 #2  0x0002ef64 in rousette::http::EventStream::enqueue (this=0xb4f22b7c, what=...) at /home/jkt/work/prog/_build/_br-cfb/build/rousette-custom/src/http/EventStream.cpp:114

The code already tried to handle this, but in signals2 the slot's
disconnection races with its execution. The usual way of solving this is
via signals2' "tracked objects", where Boost increases their refcount
while the slot is executing. We cannot use that because I *cannot*
extend the lifetime of the response by increasing a refcount somewhere
-- I just have to cease using the response object once on_close returns.

Fix this by tracking the state of the response explicitly and extending
the locking a bit. This also means that the locking was lifted from
EventStream::send_chunk() to EventStream::process().

Change-Id: If26f8152d68e770c19ff8471ebbcf8579c6829da