CI: use correct URLs for CI artifacts
When Zuul provides no artifact from a preceding job in the queue (i.e.,
there's no Depeds-on on a CzechLight/dependencies repo), this projects
needs an out-of-band method for fetching a matching artifact. Recently,
the paths for these artifacts has changed for technical reasons -- and
the changed URLs only became a problem once the changes in question were
merged. Yay.

Change-Id: Iff101660f0425149aa3e6d203443fdebf0a0a52f
Fixes: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/5608