server: get data with provided nacm user
This NETCONF-ish server will need support for authentication soon. There
will probably be a proxy server handling that and it will send a custom
HTTP header to rousette containing the name of the user.

Let's fetch the content of that header and initialize NACM. When
fetching, set NACM user to the username fetched from the header.

Change-Id: I13bdf781f3a48293a0ea2a92f1c9afc9fe40a125