CI: Switch to Fedora 38
On this system, curl is linked with the nghttp2 library. This means that
we cannot simply build a random version of libnghttp2 (with ASIO, which
means an older version, which means some unresolved symbols in `curl`,
which means no joy). Solve this by switching to the standalone
`nghttp2-asio` project.

Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/6230
Change-Id: I08e0f4f52e4cdeae6ab39baadb26ba12aa2e84f6