server: simplify listening/stopping
The underlying ASIO library has got an explicit listen_and_server()
which -- I think -- is not that much useful. Let's simplify this in the
usual RAII manner and disallow any fancy (re)starts of the server since
these are not really needed.

Actual stopping is a bit more involved because in ASIO it is not
thread-safe to call close() on a socket acceptor while its event loop
("io service") is running. That makes perfect sense, so let's make sure
that the request goes through a timer (that's the easiest option).

Change-Id: Id35e3549dbae6c8b78550456e4942da2d417e9dd
Link: https://github.com/nghttp2/nghttp2/issues/1366
Link: https://stackoverflow.com/a/17029022/2245623
Co-Authored-By: Jan Kundrát <jan.kundrat@cesnet.cz>