server: get data with provided NACM user
This NETCONF-ish server will need support for authentication soon. There
will probably be a proxy server handling that and it will send a custom
HTTP header to rousette containing the name of the user.

Let's fetch the content of that header and query data from sysrepo as
the user provided in the header.

There is one caveat and that is the anonymous user access. Before
querying as the anonymous user we must be sure that the anonymous user
has read (no write) access only to some parts of the tree. This can be
accomplished by subscribing to the changes in the `ietf-netconf-acm`
data.

Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/6367
Change-Id: I13bdf781f3a48293a0ea2a92f1c9afc9fe40a125