server: parser for data resources in URI paths
We add a parser for request URIs for data resource identifiers[1,
sec. 3.5.3]. We allow to parse path expressions for lists and leaf-lists
which is something the current code can not.

However, the key specifiers for lists and leaf-lists differ from Path
used by libyang[2]. For example, RESTCONF identifier path
`/foo:list=k1,k2` corresponds to libyang's
`/foo/list[first_key='k1'][second_key='k2']`.

This commit only parses the URI and returns slash-separated path
segments. It is not yet used. Firstly, the translation from RESTCONF
identifier paths for (leaf-)lists must be implemented.

All the structs defined in the header are a implementation detail of the
parser and the declarations are located in the "impl" header file.

[1] https://www.rfc-editor.org/rfc/rfc8040.html
[2] https://netopeer.liberouter.org/doc/libyang/master/html/howto_x_path.html

Change-Id: I5a968a83369575efbf1b213a0cd41bc327575834