server: prepare parser for request URI paths
We add a parser for request URIs for data resource identifiers[1,
sec. 3.5.3]. We allow to parse path expressions for lists and leaf-lists
which is something the current code can not.

However, the key specifiers for lists and leaf-lists differ from Path
used by libyang. For example, RESTCONF identifier path /foo:list=k1,k2
would correspond to libyang /foo/list[first_key='k1'][second_key='k2'].

This commit only parses the URI and returns slash-separated path
segments. It is not yet used. Firstly, the translation from RESTCONF
identifier paths for (leaf-)lists must be implemented.

All the structs defined in the header are a implementation detail (the
declarations are located in impl header file) of the parser.

[1] https://www.rfc-editor.org/rfc/rfc8040.html

Change-Id: I5a968a83369575efbf1b213a0cd41bc327575834