server: construct libyang paths from uris
The RESTCONF data resource identifiers [1, sec 3.5] and libyang paths
[2] differ a bit. This patch translates the former into the later.
This is needed so we can use libyang/sysrepo to perform the data
manipulation.

The RFC also states what the data resource is [1, sec. 3.5]. For
instance, you cannot query a whole list. So, some checks are needed when
translating the path.

[1] https://www.rfc-editor.org/rfc/rfc8040.html#section-3.5.3.1
[2] https://netopeer.liberouter.org/doc/libyang/master/html/howto_x_path.html

Change-Id: I541895e612860e4ba37329f7bab377e372aaa111