server: use new parser instead of the module whitelisting
So far the code relied on matching the URIs with regexps. It tried to
match the URI path with regexps that were (almost) correct and parse the
module from that and the path to nodes.
The module was needed because we only whitelisted a few YANG modules and
disabled access to any else.

This is no longer needed: We have NACM now. And we can parse URIs
conforming to the specification in RFC 8040 [1].

The test http-server is no longer needed as the functionality is removed
and replaced by our new parser.
The paths <module>:* in the http-communication tests are no longer
tested. The new parser does not support such paths because RFC 8040 [1]
does specify that such paths are valid.

[1] https://www.rfc-editor.org/rfc/rfc8040.html

Change-Id: Ib22997e7fdb535d6b0500a83dc293b9025a8da4b