server: select appropriate media types based on headers
A RESTCONF server may support both XML and JSON data formats [1]. So far
we hardcoded JSON message format everywhere but this commit selects
appropriate format based on the value of the HTTP header "accept". If
the accept header is not provided we fallback to content-type.
If the content-type is not supported, we return HTTP 415.

[1] https://www.rfc-editor.org/rfc/rfc8040.html#section-5.2

Change-Id: If656996ce0b345ff736787d1fcff6189770174a6