tests: rely on doctest+trompeloeil integration from upstream
There's been a behavior change in doctest 2.8.9 which changed how
strings are printed. This was fixed in Trompeloeil v43. Unfortunately,
we were bundling that one trivial file that sends error info from
Trompeloeil to Doctest, and that means that we were not getting the fix
from Trompeloeil.

Why did we bundle this? That's an excellent question. It goes back to
cla-sysrepo and pre-doctest days. It appears to have been introduced
back when we switched from Catch v1 to Catch v2 in cla-sysrepo's commit
59cc09af22899579a354c2ab1c6ae2f1ae40ac2d. That in itself was just a
modification to something from cla-sysrepo commit
321a3424a179e41df7bb43d101cf853993ef1c00 which refers to the upstream
wiki. Chances are that they were not providing the
`$UnitTestFramework/trompeloeil.hpp` file, so we might have had to
copy-paste like that.

Seven years later, let's just make our lives easier.

Bug: https://github.com/rollbear/trompeloeil/issues/279
Change-Id: I926951e106c8b2aa4eb5eb8e3481e046021ddcde
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/6530