PAM authentication
After a few back-and-forth attempts (such as trying to teach
nghttp2-asio how to talk over a Unix socket, or playing out with
nghttpx' mruby support for PAM) I decided to implement the
authentication within Rousette; I mainly wanted to avoid dealing with
yet another reverse proxy to delegate authentication to. As a result,
Rousette now has native linux-pam integration (and a new requirement).
The actual PAM-talking code was taken from Dovecot (and I rewrote it in
modern C++, so all bugs are mine).

If the standard HTTP header for basic auth is provided, it's checked
against the system configuration. If it's an anonymous request *and* if
anonymous requests are properly configured, it's handled as such.

Testing is done via a custom PAM configuration. There's a Berkley DB
authentication backend that ships with the upstream PAM, so I used this
pam_userdb thingy. However, since I had no clue how the BDB storage
works, I used this Python script for generating it first (insert more
passwords below):

 import bsddb3
 db=bsddb3.db.DB()
 db.open('/full/path/to/tests/pam/database.db', bsddb3.db.DB_HASH, bsddb3.db.DB_THREAD | bsddb3.db.DB_CREATE)
 db.put(b'root', b'sekrit')
 db.put(b'dwdm', b'DWDM')
 db.put(b'yangnobody', b'no chance')
 db.put(b'Aladdin', b'open sesame')

...and then dumped it through a `db_dump -p tests/pam/database.db`. I
guess I could have just as well used the `-T` option to `db_load`. Maybe
next time.

PAM apparently hard-requires a file named `other` in the specified
directory; it's probably taken as a last-resort auth method. I would
have preferred to ship a file called `tests/pam/rousette`, but hey,
`other` it is.

Change-Id: I656d40a7ff34d11b3886afdb7f4ae47deeef910b
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/br2-external/+/6495