tests: handle connection timeouts
In case of unexpected uncatched exceptions in asynchronous callbacks the
connection seems to unexpectedly terminate with the client still waiting
for the response. Let's set timeout of 3 seconds which seems reasonable
now.

Also, when doctest's FAIL is called from client.on_error callback we
obtain a bad_weak_ptr exception. Apparently it is OK if we just store
the error and call FAIL from the main thread.

Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>
Change-Id: Iad2178e415989a77070917e5a692c1a9dc7c26ab