tests: PAM: switch to pam_wrapper from pam_userdb
When I selected pam_userdb for testing, I choose it because it is a part
of upstream linux-PAM. However, distros such as Arch do not ship it by
default, and Fedora 40 will configure that module with a different
database backend (and that would require a different DB format).

It turns out that there's another mocking library for PAM, the
`pam_wrap` from `cwrap`, and its `pam_matrix` module. It has a nice
advantage that it doesn't require any fancy PAM overrides, it's just
something that gets activated by `LD_PRELOAD`. On the other hand,
pam_matrix doesn't appear to support colons in passwords, so there's one
less things that we can test.

On the plus side, it uncovered some fancy PAM bits:

- there's a `verbose` option which enables us to test the
  `PAM_ERROR_MSG` and `PAM_TEXT_INFO` branches (yay!)
- as a result, it calls the conversation function twice, the second time
  with zero number of messages, which triggered a UB in my code before

Also, now there's no need to mess with extra PAM config directories.

The upstream docs need changing because they talk about
`PAM_WRAPPER_DISABLE_DEEPBIND` environment variable whereas the actual
variable is called `UID_WRAPPER_DISABLE_DEEPBIND` in the C source code.
I'll send a patch upstream to clarify that.

Finally, it turned out that `pam_userdb` is suuuuuper sloooow on Fedora
38. A successful authentication takes about 2s, which is just
unacceptable. Note that this is definitely not the "PAM failure delay";
it's in the successful case as well.

Change-Id: Iffc7baff5b1772003f321cec2be6073857a4b502
Fixes: 542a8bd2f67fd6d77a5dcaab7bd810191e15721b PAM authentication