async sleep upon a PAM failure
IMHO the cleanest way to implement this is to let the "low-level" PAM
code just pass the requested timeout to the upper layer, and implement
the delay in there -- hence that "weird" exception.

The PAM stack adds a significant level of randomization to the actual
sleep time; I've actually seen a test failure when the "minimal expected
duration of a sleep" was set to just 2s.

In the actual server code, we have to be careful about object lifetimes
because this is C++, and we're capturing references in lambdas (yay).
The timer fires upon the timeout obtained from PAM and sends a response
back to the client, but if the client has already closed the connection
by that time, the timer gets cancelled and its "failure to fire" means
that nothing will be sent to the client. AFAICT there's no cleaner way;
we have to keep capturing the req/res variables by references because
these are non-copyable, and it is not possible to extend their
respective lifetimes anyway because they are managed by the nghttp2-asio
library.

Change-Id: I0881f004b0188c230d9e9b65e075625d7c722b98