tests: rework sysrepo init/cleanup
We used to call out to sysrepo to uninstall some modules. However, doing
that breaks down when for example there's a NACM module which refers to
a module that we're trying to uninstall. Also, the SHM files were left
all over the place. That was super fragile: on the next run, sysrepo
would choke on the inconsistency between "some state" in SHM, and a
more-or-less-cleaned, but also maybe-half-initialized state in the
repository (because module installation was done over multiple
`sysrepoctl --install` invocations).

Since we're only interested in "cleaning up", there's no need to talk to
sysrepo "nicely", and it's much easier and faster just to take out a big
hammer and start poking around in the filesystem. But since I'm worried
about a potentially unintended `rm -rf /`, I wanted to handle this via
CMake and only ever ask for file removal based on paths which are safe
to nuke.

For installation, I think it's not needed to invent a custom DSL. Let's
just build a one long lie to pass to `sysrepoctl`; that tool can do
everything that this macro needs, and then some.

Change-Id: Iac670b37f33cb31cec26a157405da48c0c491c02