tests: rework sysrepo cleanup
We used to call out to sysrepo to uninstall some modules. However, doing
that breaks down when for example there's a NACM module which refers to
a module that we're trying to uninstall. Also, the SHM files were left
all over the place.

Since we're only interested in "cleaning up", there's no need to c=talk
to sysrepo "nicely", and it's much easier and faster just to take out a
big hammer and start poking around in the filesystem. But since I'm
worried about a potentially unintended `rm -rf /`, I wanted to handle
this via CMake and only ever ask for file removal based on paths which
are safe to nuke.

Change-Id: Iac670b37f33cb31cec26a157405da48c0c491c02