tests: split the NACM- and auth-related stuff into a separate test
This patch changes the test logic. The previous version was a bit
confusing because it interleaved a DOCTEST_SUBCASE-based control flow
with a sequence of commands. I bumped into this when experimenting with
PAM timeouts which were a bit painful to test.

Since this is a second test that might run in parallel with the other
test that also instantiates a RESTCONF server that listens on some TCP
port, make sure that the port numbers are unique (and different from the
standalone binary to make stuff easier for me).

Change-Id: I77473d599e4ff802e972279b56afee27975d367e