tests: PAM: run with a separate /tmp
Unfortunately, pam_wrap performs some black magic, and as a result
of that, all tests are racing around just 62 names below /tmp. Because
that library predates `pam_start_confdir()`, it used to copy stuff into
a temporary directory. Apart from the config, it would also copy the
real system's libpam.so in there. Before that "real" library was loaded,
this copy of libpam.so was in-place binary patched so that instead of
`/etc/pam.d`, the library would read stuff from `/tmp/pam.X` where the
`X` was getting semi-randomized. This required the strlen of the mocked
`/etc/pam.d` to not change, hence yielding just one char of randomness,
the last character in `/tmp/pam.X`.

I've actually seen a (failed) race in the CI, so let's wrap this into
more bubble wrap and make sure that the test runs with a fully separate
/tmp. Doing that requires a custom UID namespace and faux UID 0 in there
as well.

Since I'm on NixOS with very funny permissions on `mount`, this also
needs a different CMP0109 policy, hence the bump of the minimal reuired
version.

Finally, we only need the real test binary to run under this magic env,
so the original `wrap_test_pam` macro really has to be rolled into the
shell invocation.

The TEST_COMMAND syntax is only used by velia (another project of ours
that we're regularly copying stuff from), and in there it's used for a
custom network namespace. Let's preserve that option so that we can
(eventually) sync these scripts to other projects as well.

Change-Id: I8084724257c5b9eb1dd6a6ba2fa1cccf0008d7ce