tests: add NACM rule for anonymous user to read example module
We are going to use this module for writing tests in the next commit.
I wanted this small change to stand out because the output of one
reading test is different now as some sysrepo data are now found.

I have spent some time on this before I realize what is happening. That
is why I don't want to squash this in the future commits. :(

Change-Id: I55c4840ddbd8e9a081fab92e8c853fc4171a9321