server: collect type of resource from URI
RESTCONF has several resources under /restconf. We already implement
/restconf/data (RFC 8040 [1]) and /restconf/ds (RFC 8527 [2]).

In the next commit we will add also /restconf/operations for RPC calls.
However, to verify that RPCs were invoked with corrent resource we need
to remember the type of the resource. This commit introduces the basic
support as a preparation for the next commit adding RPC/Action support.

Change-Id: I6ade004d405d406569babe9d7629663fd67d05b2