restconf: collect URI prefix from parser
RESTCONF has several resources under /restconf. We already implement
/restconf/data (RFC 8040 [1]) and /restconf/ds (RFC 8527 [2]).

In the near future we will add /restconf/operations prefix for RPC calls.
However, to verify that RPCs were invoked with current resource we need
to remember the type of the resource. This commit introduces the basic
support as a preparation for the future support of RPC/action URIs.

[1] https://www.rfc-editor.org/rfc/rfc8040.html
[2] https://www.rfc-editor.org/rfc/rfc8527.html

Change-Id: I6ade004d405d406569babe9d7629663fd67d05b2