restconf: rename ResourcePath to URI
The struct ResourcePath which encapsulates parsed URI (i.e., the
datastore and path segments) may be confusing a bit. It represents the
parsed URI so why not call it just URI?

Yeah, it is not exactly "URI" we are missing schema and other stuff, but
we think this name represents the purpose of the struct better.

Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>
Change-Id: If1845171cff2368837c74bc9b0c61734f71a9030