restconf: URI parser now considers a HTTP method and URI prefix
In a preparation for RPC and action (in RESTCONF term just "operations")
we rewrite the URI parser a little.
The parser will now also consider URI prefix type and HTTP method to
determine if the URI is correct.
For example, if one sends a POST request to /restconf/data/mod:rpc we
already know that this URI is invalid because you can't invoke and
operation using /restconf/data prefix.

Change-Id: I095b5f77372583a2ad6fc9e88e3960bfae7f704c