restconf: reorganize URI parser node validation
Checking if every segment of the URI correspond to valid data path
worked up until now. In the future, when we implement more HTTP methods
and resources, it might get more messy.
We believe that it is enough to just check the validity of the path for
the HTTP method once: Just after we obtain the final
libyang::SchemaNode corresponding to the URI (this however does not
apply for RPC child nodes).
It should be reasonably fast to iterate all the way up in case we need
to check something in parent nodes in the future.

This commit only reorganizes internal code of the URI parser, there is
no visible external change to the parser behaviour.

Change-Id: I9bab606cbee1767cf4a95e5e2393aabe2fc6df11