restconf: URI parser schema node validation redesign
Checking if every segment of the URI correspond to valid data path
worked up until now. In the future, when we implement more HTTP methods
and resources, it might get more messy.
We believe that it is enough to just check the validity of the path for
the HTTP method once: Just after we obtain the final
libyang::SchemaNode. It should be reasonably fast to iterate all the way
up in case we need it (if we want to check that our parent is a RPC
node, etc.).

This change has positive side-effects too: When requesting input
nodes of an action/RPC we can easily output the exact path of the
input node instead of the RPC path (where we stopped before).
Also if we pass the HTTP method inside the parser we can easily reject
PUT on datastore path.

Change-Id: I9bab606cbee1767cf4a95e5e2393aabe2fc6df11