restconf: don't return datastore from asLibyangPathSplit
This function was used for splitting the path only. We need that in PUT
requests.
Originally the idea was to make a pair of functions: asLibyangPath and
asLibyangPathSplit and use the former for GET requests and the latter
for PUT requests. However, in the meantime we use the former for both
anyway. The split function is used in PUT requests to split the path but
the datastore is unused anyway.
Let's nuke that.

Change-Id: Iaa657a7dc796b4331585aa3baedf104176fadd82