restconf: move unsupported HTTP method check into URI parser
We support only GET and PUT so far. Other HTTP methods are rejected with
RESTCONF-compliant error response.

Also, our URI parser is now a function of the HTTP method too and it
will soon be refactored in the way that it will return the action that
is determined by the URI. So it makes sense to move the changed line of
code into the parser to reject (yet) unsupported HTTP requests.

Change-Id: Ice90bae5030db0e9b6e45f81cc767b8d98275568