restconf: return action to be executed from URI parser
URI parser knows about the URI more than anyone else. It also know about
HTTP request method. We can actually determine what action is to be
executed based on the HTTP method and the URI.

We will soon be implementing POST requests which can mean two different
things: Either RPC execution, or writing data into the tree. And the
actual performed operation is determined from the URI.

This commit adds an type of the request to the return value of the
parser. Server then only simply branches on the value of request type.

Change-Id: Ic06e9c2b9f35309b343f744849d7d94b86b07898