restconf: replace InvalidURIException with ErrorResponse
In 38e7f49 ("restconf: GET on operations resource should respond with
HTTP 405") we throwed ErrorResponse from the uri parser for the first
time. Actually, I think that this is a good idea, because throwing an
exception from the parser means that user provided invalid URI and we
have to reject the request.
Currently we throw InvalidURIException which holds only an information
about what happened and Server must catch the exception and set up the
error response.
However, the Server could not possibly know what went wrong and how it
should respond (well, it could but it would have to pattern match the
exception message). The URI parser knows exactly what went wrong and
what HTTP status code, error tag and type it should report.

This commit therefore removes the InvalidURIException from the codebase
and replaces its usage with ErrorResponse.

This commit does not change the actual error responses behaviour.
The responses remain the same for now. Somebody will have to go through
all the messages and check if they are correct in the future.

However, this commit changes the URI parser tests. The test expectations
for expected failures are rewritten; they now assert all the details of
the failure (so far we only checked that the parser threw).

Change-Id: Ic0e172e0e8e92c74fb59f946edc0ed8e0478a80a