restconf: PUT requests on datastore resource should throw in URI parser
So far, we reject datastore resource PUT requests (e.g.,
PUT /restconf/data) in Server.cpp when we find out that the path can
not be split with asLibyangPathSplit function.

This can rejected already in the URI parser. The parser, however, needs
to know the request HTTP method.

I'm doing this patch now, because I will be modifying more stuff in URI
handling and this non-systematical throw gets in a way.

Change-Id: I2c3824e25b3f7061fb22291e2984705b9832e1d1