restconf: error handling for RPC URIs
I wanted to redesign URI parser in a way, that we follow the schema
nodes along the URI. Then we say whether the URI is contextually valid
or not.
However, for URIs like /mod:rpc/inputLeaf or /mod:rpc/outputLeaf we
can't do that. It turns out that in libyang, you can only iterate
through input or output RPC nodes and you must select your choice in the
Context::findPath call. However, I can't possibly know whether the RPC
child node is an input or output node. Perhaps I could try both?
After all, this is just an intelectual exercise. Anything under RPC node
is invalid request anyway, so just respond with 400 status code
when we detect that user requested such URI.

This patch changes the error messages for some requests so it is clear
that user either used an invalid method with RPC node or tried to
request a child of RPC node.

Change-Id: I2dad3d777ef0f30e916f344e6b81d58ffd42156b