cmake: tests: cleanup: don't error out if there is nothing to remove
I'd like to be able to keep running this "test" (it's a cleanup code for
a test, but in CMake/CTest, that means that it's a test) no matter
whether the initialization and the actual test have run. Before this
commit, it would fail because these globs produced no results, and
therefore the CMake code would call something with too few arguments.

This relies on CMake silently ignoring nonexisting files. That's easier
than checking for list length because that would require an extra
variable, etc.

Change-Id: Iabd23b6cb98ac81150bb38dd44b5e3f01cf33e2b