cmake: tests: also clean sysrepo state before running the tests
When a test run is interrupted for some reason, the cleanup phase won't
get executed, and some (possibly corrupt) state is left in the SHM files
and in the on-disk repository. As a result, the subsequent run would
often fail. This was not a problem for the CI (which always starts from
scratch and using a 100% clean state), but it was annoying enough for
developer builds.

Fix that by introducing another fixture, the preinit one, which does the
exact same thing as the cleanup one.

Change-Id: I1a988e183ac6566823fdee7f81ce746ff931b341