restconf: parse YANG schema URIs
Next commit adds /yang endpoint that returns YANG schemas for specified
module. To add the endpoint, we must parse such URIs first.

The URIs are in the format /yang/module_name(@revision(.yang)?)?

The .yang suffix is allowed only when revision is specified because a
module name can contain dots so /yang/yang.yang could be a module named
yang with a .yang file suffix but also a module named yang.yang.

Change-Id: I57c94a59350cf692bab26aab2cd239599c657c43