restconf: add endpoint for getting YANG schemas
This is a preparation for implementing Schema Resource defined in
RESTCONF RFC [1, sec. 3.7].

This commit adds endpoint which returns YANG schemas for queried module.
The RFC states that

  Note that there is no required structure for this URL.

I have chosen a /yang/module(@revision)? endpoint for that. I did not
want to put this under the /restconf root so we do not clash with
paths defined by some possible future restconf RFCs.

[1] https://www.rfc-editor.org/rfc/rfc8040#section-3.7

Change-Id: I6a58fd88a33209b13dc5dc0f96b06993d890daed