restconf: show schema locations for modules in ietf-yang-library
The ietf-yang-library module [1] contains a list of all modules the
server knows about.
It can also point to a file from which the schema was loaded from.
Sysrepo takes care of filling the lists and locations to the files.

The RESTCONF server should, however, not point to internal filepaths
but rather point to some HTTP resource where one can download the file.

This commit implements it by overriding the values using sysrepo
operational pull callbacks where we overwrite the local values.

The server, however, also runs behind a reverse proxy and it does not
know about the original requests' Host header. Until we solve how to
deal with this we omit the schema and host part from the URL.

[1] https://datatracker.ietf.org/doc/html/rfc8525

Change-Id: I7c3e4d8c92c01da0f472bcfa2573dffb3cbc85cd