restconf: show schema locations for modules in ietf-yang-library
The ietf-yang-library module [1] contains a list of all modules the
server knows about.
It can also point to a file from which the schema was loaded from.
Sysrepo takes care of filling the lists and locations to the files.

The RESTCONF server should, however, not point to internal filepaths
but rather point to a HTTP resource where users can download the file.
This commit implements it by replacing the location values in the YANG
tree.

We, however, have to provide full URLs. Our server runs behind a reverse
proxy. We try to parse the forwarded HTTP header and guess the protocol
and host from its value. In case we can't parse both values then the
locations leafs are not provided.

[1] https://datatracker.ietf.org/doc/html/rfc8525

Change-Id: I7c3e4d8c92c01da0f472bcfa2573dffb3cbc85cd