restconf: authorize for GET on /yang/ endpoint
We want to limit fetching yang schemas for some users. The point is that
users can scan for "known vulnerabilities" by checking
/restconf/data/ietf-yang-library:* endpoints and then fetch the schemas
at the /yang endpoint.

The RESTCONF already uses NACM to some level so we thought that we might
use it for this as well. Operators can hide
/ietf-yang-library:yang-library contents using NACM (see previous
commit) but that doesn't limit the /yang endpoint access. The users can
still "guess" the module names and revisions.

This commit implements NACM authorization at /yang endpoint as well.
However, the /yang endpoint is not anywhere within our data tree so one
can not set NACM rules for it directly. Instead, we check that the
authorized user can read the module entry in
/ietf-yang-library:yang-library/module-set/module list. If so, then we
provide the yang schema. Otherwise, it is 404.

Change-Id: I689ba13e61fa6b9f0ba617d1ea3e7b32c7ad3b6d
Co-authored-by: Jan Kundrát <jan.kundrat@cesnet.cz>