restconf: allow fetching submodules
In 2e19ce3c ("restconf: add endpoint for getting YANG schemas") we added
a support for fetching YANG modules via HTTP. However, this approach
does not work for submodules, because they could not be retrieved using
libyang's Context::getModule function. There was actually no function in
libyang-cpp, now we have it, so let's use it.

Fixes: 2e19ce3c9923792ac28c7549d121476b7b595f93
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/6997
Change-Id: I4d9964ffb678b3a63d13a9a9d01fc42eaa015945