tests: check that we replace all yang-library location leaves
In 9ff26d45b ("restconf: show schema locations for modules in
ietf-yang-library") we started rewriting location leafs in
ietf-yang-library when the data are requested via RESTCONF. This commit
only adds test that we remove all the location leaf-list nodes. Sysrepo
is (currently) inserting only one entry in the leaf-list so we did not
test this case.

Change-Id: I9458a46336bb8237cd07730ea4aafb49fb093bf2