tests: move URI testing for RPC nodes into their own section
There is a large section of tests in the URI parser tests that tests for
invalid data resources for GET, PUT, and DELETE methods. It also
contains a section where we check that these methods don't work with RPC
and action nodes.
If we take the section for RPC and action nodes away, this whole section
can test POST requests for creating data too.

This will come handy in the next commit where we add a POST for creating
data. We will be able to reuse these tests for POST method as well.

Change-Id: Icdc28a708e713dda97bddb733c500a88a3169e82