restconf: rename URI to URIPath
We are going to implement query parameters soon. It involves parsing
another part of the uri, the querystring. This structure, which
contains the URI path split by delimiters, really contains only the
path, not the query and other parts of the URI.

Renaming it to URIPath might clarify it more clearly than just stating
that in the doxygen comment.

Change-Id: I945c718df7e23556c3268f76436eb7c15791a484