tests: test explicit default handling basic mode
RESTCONF RFC [1] says that server should report the mode of handling
default values. We have not implemented capabilities list yet but we
should choose how the default values should be handled. This affects GET
operations that do not specify with-defaults query parameter, and create
operations on nodes with default values. These modes are defined in a
separate RFC that deals with this problem in NETCONF [2].

We rely on libyang+sysrepo for the implementation. By default they work
with explicit default data mode so there should be no more extra work to
be done from us.

This commit adds some test to check that we really behave as the RFC
specifies.
We are unsure about that one test with FIXME comment, there is some
discussion in sysrepo's bugtracker [3].

[1] https://datatracker.ietf.org/doc/html/rfc8040#section-3.5.4
[2] https://datatracker.ietf.org/doc/html/rfc6243#section-2
[3] https://github.com/sysrepo/sysrepo/issues/3283

Change-Id: I5548f65463b02c8919323e12b6bfd36d2b321526