restconf: handling of default values in leafs
Restconf RFC [1] says that server should report the mode of handling
default values. We have not implemented capabilities list yet but we
should choose how the default values should be handled.

I argue that reporting all default values is most transparent to
clients. They do not have to check the model to see if the leaf value
is missing in the output because it was set to the default value.

[1] https://www.rfc-editor.org/rfc/rfc8040.html#section-3.5.4

Change-Id: I5548f65463b02c8919323e12b6bfd36d2b321526